refactor: split skia text replay
