+++ src/renderer/skia/mod.rs
@@ -5,5 +5,6 @@
 mod equation_conv;
 mod image_conv;
 mod renderer;
+mod text_replay;
 
 pub use renderer::SkiaLayerRenderer;
